Refactor rendering in neural evolution example
Code responsible for rendering Creatures, Berry Bushes and World now are
methods of their respective classes, and not in Renderer.
Renderer only contains code for drawing simple shapes such as circles and
rectangles and also text.

CSVFile handling, which is not rendering but has been put in the
Renderer class has also been refactored out of the Renderer class and
put into World class.