Add check for overfitting(probably) for examples/backpropagation
This work by evaluating the trained neural network both on the training
and testing data set, and if there were to be overfitting, the accuracy
of training data set should be much higher than that of the testing data
set. At least currently, that does not seem to be the case.