Improve implementation of Model class
boost::graph is used in the implementation of Model class.
Previously, both data(input/output/gradients) and layers are stored as
vertex of graph. The new implementation stores layers as edges of the
graph, and data as vertices of the graph. This is to prepare for the
implementation of feedback connection.