Change API for Model::{feedForward,backPropagate}()
Model::feedForward() no longer return output directly.
To get the output for Model::feedForward(), call Model::output().
This make the API consistent with Model::backPropagate(), and make it
possible to retrieve the output repeatedly.
The change is possible following the previous commit c787a5 that changes
the usage of graph such that the output of Model is stored in a
dedicated class Vertex vertex of boost::graph.
Model::backPropagate() also no longer required output to be passed
separately, since output of the Model is now stored internally.