Expose the full boost graph API in Model
A boost graph with vertices of Node type and edges
of Connection type is passed to Model constructor to construct the Model
instead of the old addNode, addConnection and build member function
call.

This is because no abstraction is better than leaky abstraction, and
having abstraction would disallow the user to take full advantage of the
API of boost graph library.